requirements.txt
requests
python-dotenv
youtube-transcript-api
